Relocate Circulars Archive header into header file, rename files to better distinguish from individual circular routes (#3747)
* Rename archive files for clarity

* Relocate header content into header file

* Remove index from children of header

* Remove duplicate moderator components from route

* Fix location of state vals
