--- app/routes/circulars._archive._index/CircularsHeader.tsx
@@ -1,30 +1,0 @@
-/*!
- * Copyright © 2023 United States Government as represented by the
- * Administrator of the National Aeronautics and Space Administration.
- * All Rights Reserved.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-import { Link } from '@remix-run/react'
-
-export default function () {
-  return (
-    <>
-      <h1>GCN Circulars</h1>
-      <p className="usa-paragraph">
-        <b>
-          GCN Circulars are rapid astronomical bulletins submitted by and
-          distributed to community members worldwide.
-        </b>{' '}
-        They are used to share discoveries, observations, quantitative near-term
-        predictions, requests for follow-up observations, or future observing
-        plans related to high-energy, multi-messenger, and variable or transient
-        astrophysical events. See the{' '}
-        <Link className="usa-link" to="/docs/circulars">
-          documentation
-        </Link>{' '}
-        for help with subscribing to or submitting Circulars.
-      </p>
-    </>
-  )
-}
